Move TestRoboelectric class to base-android module
