+++ base-android/src/test/kotlin/com/jpaya/base/ui/bindings/ImageViewBindingsTest.kt
@@ -17,8 +17,7 @@
 package com.jpaya.base.ui.bindings
 
 import android.widget.ImageView
-import com.jpaya.base.R
-import com.jpaya.libraries.testutils.robolectric.TestRobolectric
+import com.jpaya.base.testutils.TestRobolectric
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Before
