+++ base-android/src/test/kotlin/com/jpaya/base/ui/extensions/RecyclerViewExtensionsTest.kt
@@ -19,7 +19,7 @@ package com.jpaya.base.ui.extensions
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.jpaya.libraries.testutils.robolectric.TestRobolectric
+import com.jpaya.base.testutils.TestRobolectric
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.instanceOf
 import org.junit.Assert.assertNotNull
