+++ buildSrc/src/main/kotlin/dependencies/Libraries.kt
@@ -172,16 +172,16 @@ object TestLibraries {
 
     private const val ARCH_CORE = "androidx.arch.core:core-testing:${Versions.ARCH_CORE}"
     private const val ASSERTJ = "org.assertj:assertj-core:${Versions.ASSERTJ}"
-    private const val CORE = "androidx.test:core:${Versions.TEST}"
+    const val CORE = "androidx.test:core:${Versions.TEST}"
     private const val COROUTINES_TEST = "org.jetbrains.kotlinx:kotlinx-coroutines-test:${Versions.COROUTINES}"
     private const val EXT = "androidx.test.ext:junit:${Versions.EXT}"
     private const val FRAGMENT_TEST = "androidx.fragment:fragment-testing:${Versions.FRAGMENT}"
     private const val HAMCREST = "org.hamcrest:hamcrest:${Versions.HAMCREST}"
-    private const val JUNIT = "junit:junit:${Versions.JUNIT}"
+    const val JUNIT = "junit:junit:${Versions.JUNIT}"
     private const val MOCKITO = "com.nhaarman.mockitokotlin2:mockito-kotlin:${Versions.MOCKITO}"
     private const val MOCKK = "io.mockk:mockk:${Versions.MOCKK}"
     private const val RAINBOW_CAKE = "co.zsmb:rainbow-cake-test:${Versions.RAINBOW_CAKE}"
-    private const val ROBOELECTRIC = "org.robolectric:robolectric:${Versions.ROBOELECTRIC}"
+    const val ROBOELECTRIC = "org.robolectric:robolectric:${Versions.ROBOELECTRIC}"
     private const val ROOM = "androidx.room:room-testing:${Versions.ROOM}"
     private const val RULES = "androidx.test:rules:${Versions.TEST}"
     private const val RUNNER = "androidx.test:runner:${Versions.TEST}"
